+++ TRABALHO ALGORITMOS/bubble.h
@@ -2,6 +2,6 @@
 #define H_ 
 #include <windows.h>
 
-void BubbleSort(int *numeros,int TAM100,int *COMPARACAOGLOBAL,int *MOVIMENTACAOGLOBAL);
+void BubbleSort(int *numeros,int TAM100,unsigned long int *COMPARACAOGLOBAL,long int *MOVIMENTACAOGLOBAL);
 
 #endif
