+++ TRABALHO ALGORITMOS/quick.h
@@ -2,6 +2,6 @@
 #define H_ 
 #include <windows.h>
 
-void QuickSort(int *numeros,int TAM100,int *COMPARACAOGLOBAL,int *MOVIMENTACAOGLOBAL);
+void QuickSort(int *numeros,int TAM100,unsigned long int *COMPARACAOGLOBAL,unsigned long int *MOVIMENTACAOGLOBAL);
 
 #endif
